+++ CHANGELOG.md
@@ -9,29 +9,52 @@ The `Unreleased` section name is replaced by the expected version of next releas
 ## [Unreleased]
 
 ### Added
-
-- now targets `Microsoft.Azure.Cosmos` v `3.9.0` (instead of `Microsoft.Azure.DocumentDB`[`.Core`] v 2.x) [#144](https://github.com/jet/equinox/pull/144)
-
 ### Changed
 
-- Update to `3.1.101` SDK
+- target `Microsoft.Azure.Cosmos` v `3.9.0` (instead of `Microsoft.Azure.DocumentDB`[`.Core`] v 2.x) [#144](https://github.com/jet/equinox/pull/144)
+- target `EventStore.Client` v `20.6` (instead of v `5.0.x`) [#224](https://github.com/jet/equinox/pull/224)
 - Retarget `netcoreapp2.1` apps to `netcoreapp3.1` with `SystemTextJson`
 - Retarget Todobackend to `aspnetcore` v `3.1`
 - Target `FSharp.Control.AsyncSeq` v `2.0.23`
-- Updated AzDO CI/CD to use `windows-latest`
+- `Equinox.Tool`: Target `FSharp.Core` v `4.7.1`
+- Update AzDO CI/CD to use `windows-latest`
+- Update to `3.1.101` SDK
 - Remove `module Commands` convention from in examples
 - Revise semantics of Cart Sample Command handling
 - `Cosmos:` Removed [warmup call](https://github.com/Azure/azure-cosmos-dotnet-v3/issues/1436)
+- Simplify `AsyncCacheCell` [#229](https://github.com/jet/equinox/pull/229)
 
 ### Removed
 ### Fixed
+
+<a name="2.3.0-rc1"></a>
+## [2.3.0-rc1] - 2020-08-31
+
+### Added
+
+- `Cosmos`: `Prune` API to delete events from the head of a stream [#233](https://github.com/jet/equinox/pull/233)
+
+### Changed
+
+- `Cosmos`: Tweaked log messages
+
+<a name="2.2.0"></a>
+## [2.2.0] - 2020-08-04
+
+### Added
+
+- Add `AsyncBatchingGate` [#228](https://github.com/jet/equinox/pull/228)
+
+### Fixed
+
+- `EventStore`: Pin `EventStore.Client` to `[5.0.1,6)` to avoid conflicting changes in `EventStore.Client` v `20.06` [#223](https://github.com/jet/equinox/pull/223)
 
 <a name="2.1.0"></a>
 ## [2.1.0] - 2020-05-22
 
 ### Added
 
-- Add `eqx dump -b`, enabling overriding of Max Events per Batch
+- `eqx dump -b`, enabling overriding of Max Events per Batch
 - `MemoryStore`: Add `Committed` event to enable simulating Change Feeds in integration tests re [#205](https://github.com/jet/equinox/issues/205) [#221](https://github.com/jet/equinox/pull/221)
 
 ### Changed
@@ -404,7 +427,9 @@ The `Unreleased` section name is replaced by the expected version of next releas
 
 (For information pertaining to earlier releases, see release notes in https://github.com/jet/equinox/releases and/or can someone please add it!)
 
-[Unreleased]: https://github.com/jet/equinox/compare/2.1.0...HEAD
+[Unreleased]: https://github.com/jet/equinox/compare/2.3.0-rc1...HEAD
+[2.3.0-rc1]: https://github.com/jet/equinox/compare/2.2.0...2.3.0-rc1
+[2.2.0]: https://github.com/jet/equinox/compare/2.1.0...2.2.0
 [2.1.0]: https://github.com/jet/equinox/compare/2.0.2...2.1.0
 [2.0.2]: https://github.com/jet/equinox/compare/2.0.1...2.0.2
 [2.0.1]: https://github.com/jet/equinox/compare/2.0.0...2.0.1